refactor: rename Checkbox to InputCheckbox
